requirements.txt

flet
fastapi
uvicorn
openai
requests
python-multipart
pandas
openpyxl
pytesseract
pillow
